requirements.txt
torch>=1.13.0
torchvision>=0.14.0
gradio>=3.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0

transformers>=4.30.0
torch>=1.13.0
Pillow>=9.0.0
